[WIP] added HCTR
Not quite ready for publishing.

* [ ] Should be generic over the block cipher and AXU universal hash
function.

* [ ] Documentation... Not all AXU universal hash functions are valid;
there are at least two additional properties that must hold.

* [ ] Is anything else missing?

* [ ] Publish to crates.io under RustCrypto

Why is Aes + ctr about half the performance of aesni w\ the ctr feature?

Please confirm this crate belongs here.

HCTR is an SPRP but it builds on a block cipher and an AXU universal
hash function, it is a construction or cipher family; not a cipher
itself.