aes: remove Copy from public types of "soft" impls (#215)
When using the `autodetect` module, we store the possible backing AES
implementations in a `union`, which presently requires all fields be
`Copy` until `ManuallyDrop` stabilizes in Rust 1.49.

However, the "soft" types are directly exposed on non-x86 architectures
or when the `force-soft` feature is enabled. Having these types be
`Copy` when the autodetect wrappers aren't is inconsistent, and it's
better to require an explicit clone for these to make users think when
they make copies of material.

As a workaround until `ManuallyDrop` lands, this gates the `Copy` impl
on the "soft" types to only cases where it is hidden behind the
autodetect facade (i.e. when they presently *need* to be `Copy`).