aes: replace AES-NI macros with inline(always) functions
The previous usage of macros made refactoring extremely difficult, with
small changes resulting in hundrdeds upon hundreds of errors when the
macros are expanded, as opposed to a single error in a particular
function.

This commit replaces the AES-NI `util.rs` macros with
`#[inline(always)]` functions.

There is no change to the generated assembly.