aes: rework backends (#442)
This PR unifies code between AES-NI and ARM backends and prepares ground
for future removal of duplicated definitions of AES types in
`autodetect`, `soft`, `ni`, and `armv8` modules. Additionally, it allows
to quickly change number of blocks processed in parallel by different
intrinsics-based backends instead of hardcoding it to 8 blocks.